Add trailing newline to files without one.
Several of the files in https://github.com/chrismiller/TeensyEyes do not have
final newlines in them.  Several of the programs that process text and
differences that run under Linux do not work well with a file that does not
have a final newline.  This patch adds a final newline to the appropriate
files.

2022-11-27  Michael Meissner  <github@the-meissners.org>

	* config.h: Add trailing newline.
	* displays/Display.h: Likewise.
	* displays/GC9A01A_Display.h: Likewise.
	* eyes/eyes.h: Likewise.
	* main.cpp: Likewise.