Allow config.h to change some parameters.
At the moment, you need to edit main.cpp to change some of the defaults.  This
patch allows config.h to override some of the defaults using macros.  If the
macro is not set, the default value is used.  Since the code uses constexpr for
these values, I decided to use macros to override the defaults, but to use
those macros to set up the constexpr.

2022-11-26  Michael Meissner  <github@the-meissners.org>

	* config.h (CONFIG_EYE_DURATION_MS): Add macro in comments, which
	config.h could use to override the defaults.
	(CONFIG_BLINK_PIN): Likewise.
	(CONFIG_JOYSTICK_X_PIN): Likewise.
	(CONFIG_JOYSTICK_Y_PIN): Likewise.
	(CONFIG_LIGHT_PIN): Likewise.
	* main.cpp (CONFIG_EYE_DURATION_MS): New macro to allow config.h to
	override the defaults.
	(CONFIG_BLINK_PIN): Likewise.
	(CONFIG_JOYSTICK_X_PIN): Likewise.
	(CONFIG_JOYSTICK_Y_PIN): Likewise.
	(CONFIG_LIGHT_PIN): Likewise.
	(EYE_DURATION_MS): Use the macro for the defaults.
	(BLINK_PIN): Likewise.
	(JOYSTICK_X_PIN): Likewise.
	(JOYSTICK_Y_PIN): Likewise.
	(LIGHT_PIN): Likewise.