After first iteration, pick eye pattern randomly.
This patch changes the order eyes are used.  Previously, we always went in
order, and when we got to the end, we started back at the beginning.

With this patch, it still does the eyes in order for the first pass, but after
the first pass it picks the next eye at random.  When it picks an eye at
random, it makes sure we don't pick the same eye.

2022-11-27  Michael Meissner  <github@the-meissners.org>

	* main.cpp (nextEye): After going through the first set of eyes
	sequentially, pick the next eye at random.